Twitter_Post/requirements.txt
pandas
httpx
pysocks
xlsxwriter
openpyxl